refactor(orgchart): relocate HTTP adapter to Infrastructure folder
Move OrgChartService concrete class from Features/OrgChart/Services/
to Features/OrgChart/Infrastructure/ to match the Services-vs-Infrastructure
boundary documented in docs/architecture/filesystem.md and used by every
other feature module. IOrgChartService interface stays in Services/ as
the abstraction the use-case handler depends on. Module gains one using
directive to resolve the relocated concrete class for the existing
AddHttpClient<IOrgChartService, OrgChartService>() registration.

No behavior change, no API change, no test change.
